Добавлены тесты. Долгоиграющий тест генерации множества случайных чисел не вызывается в отладочной конфигурации. Ускорен агоритм генерации множества случайных чисел.

+++ PrimeNumbersSetGenerator/PrimeNumbers.h
@@ -1,5 +1,5 @@
 #include "stdafx.h"
 
 bool IsValidUpperBound(int upperBound);
-std::set<int> GeneratePrimeNumbersSet(size_t upperBound);
+std::set<int> GeneratePrimeNumbersSet(int upperBound);
 void PrintPrimeNumbersToOutput(std::set<int> const &primeNumbers);
